pyproject.toml
[tool.poetry]
name = "devsearch-practice"
version = "0.1.0"
description = "Django course practice"
authors = ["mottodor"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
Pillow = "^9.0.1"
whitenoise = "^6.0.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
